requirements.txt
# ACCT445-Showcase Dependencies
# Lightweight - no large files required

# Core data manipulation
pandas>=2.1.0
numpy>=1.26.0

# Market data
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Text analysis and readability metrics
textstat>=0.7.3
nltk>=3.8

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0
linearmodels>=5.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
streamlit>=1.28.0

# Documentation
sphinx>=8.2.0
sphinx-rtd-theme>=3.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
tqdm>=4.66.0
joblib>=1.3.0
toml>=0.10.2
schedule>=1.2.0
dvc>=3.0.0
pre-commit>=3.5.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Code quality (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0

# Production scheduler and authentication
apscheduler>=3.10.0
filelock>=3.12.0
pytz>=2023.3
streamlit-authenticator>=0.2.3
pyyaml>=6.0
